https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved jmap-client-wrapper services to esn.inbox.libs
