+++ src/linagora.esn.unifiedinbox/app/services.js
@@ -2,7 +2,7 @@ const _ = require('lodash');
 
 require('./services/with-jmap-client/with-jmap-client.js');
 require('./services/config/config.js');
-require('./services/jmap-client-wrapper/jmap-client-wrapper.service.js');
+require('../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('./services/jmap-helper/jmap-helper.js');
 require('./services/mailboxes/mailboxes-service.js');
 require('./services/hook/email-sending-hook.service.js');
