+++ src/linagora.esn.unifiedinbox/app/services/attachment-jmap/attachment-jmap.service.js
@@ -1,5 +1,5 @@
 require('../../services.js');
-require('../../services/jmap-client-wrapper/jmap-client-wrapper.service.js');
+require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('../../services/attachment-jmap/attachment-jmap.constants.js');
 
 (function(angular) {
