--- src/linagora.esn.unifiedinbox/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js
@@ -1,28 +1,0 @@
-(function(angular) {
-  'use strict';
-
-  angular.module('linagora.esn.unifiedinbox')
-
-    .factory('jmapDraft', function($window) {
-      return $window.jmapDraft;
-    })
-
-    .factory('dollarQPromiseProvider', function($q) {
-      return {
-        newPromise: function(resolver) {
-          return $q(resolver);
-        }
-      };
-    })
-
-    .factory('dollarHttpTransport', function($http) {
-      return {
-        post: function(url, headers, data) {
-          return $http.post(url, data, { headers: headers }).then(function(response) {
-            return response.data;
-          });
-        }
-      };
-    });
-
-})(angular);
