+++ src/linagora.esn.unifiedinbox/app/services/mailboxes-filter/mailboxes-filter-service.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 
 require('../mailboxes/mailboxes-service.js');
 require('./mailboxes-filter-service.constants.js');
-require('../jmap-client-wrapper/jmap-client-wrapper.service.js');
+require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('../../services.js');
 require('../with-jmap-client/with-jmap-client.js');
