+++ src/linagora.esn.unifiedinbox/app/services/mailboxes/mailboxes-service.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 require('../with-jmap-client/with-jmap-client.js');
-require('../jmap-client-wrapper/jmap-client-wrapper.service.js');
+require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('../../services.js');
 require('./special-mailboxes.js');
 require('./shared-mailboxes.js');
