+++ src/linagora.esn.unifiedinbox/app/services/models/emailer.run.js
@@ -1,4 +1,4 @@
-require('../jmap-client-wrapper/jmap-client-wrapper.service.js');
+require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('./emailer.service.js');
 
 (function(angular) {
