+++ src/linagora.esn.unifiedinbox/app/services/models/mailbox.run.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-require('../jmap-client-wrapper/jmap-client-wrapper.service.js');
+require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('../../services.js');
 
 (function(angular) {
